Fix compilation error
Fix by updating local align variable from int to long so as not to lose
precision.

Compilation error:
ext/murmur/murmur_hash.cpp:89: error: cast from ‘const unsigned char*’ to
‘int’ loses precision

+++ ext/murmur/murmur_hash.cpp
@@ -86,7 +86,7 @@ unsigned int MurmurHashAligned2 ( const void * key, int len, unsigned int seed )
 
   unsigned int h = seed ^ len;
 
-  int align = (int)data & 3;
+  long align = (long)data & 3;
 
   if(align && (len >= 4))
     {
